# Remove unsafe equalStr