--- a/bytes_safe.go
+++ b/bytes_safe.go
@@ -1,3 +1,4 @@
+//go:build appengine || appenginevm
 // +build appengine appenginevm
 
 package jsonparser
@@ -8,8 +9,8 @@ import (
 
 // See fastbytes_unsafe.go for explanation on why *[]byte is used (signatures must be consistent with those in that file)
 
-func equalStr(b *[]byte, s string) bool {
-	return string(*b) == s
+func equalStr(b []byte, s string) bool {
+	return string(b) == s
 }
 
 func parseFloat(b *[]byte) (float64, error) {