--- a/bytes_unsafe.go
+++ b/bytes_unsafe.go
@@ -1,25 +1,17 @@
+//go:build !appengine && !appenginevm
 // +build !appengine,!appenginevm
 
 package jsonparser
 
 import (
 	"reflect"
+	"runtime"
 	"strconv"
 	"unsafe"
-	"runtime"
 )
 
-//
-// The reason for using *[]byte rather than []byte in parameters is an optimization. As of Go 1.6,
-// the compiler cannot perfectly inline the function when using a non-pointer slice. That is,
-// the non-pointer []byte parameter version is slower than if its function body is manually
-// inlined, whereas the pointer []byte version is equally fast to the manually inlined
-// version. Instruction count in assembly taken from "go tool compile" confirms this difference.
-//
-// TODO: Remove hack after Go 1.7 release
-//
-func equalStr(b *[]byte, s string) bool {
-	return *(*string)(unsafe.Pointer(b)) == s
+func equalStr(b []byte, s string) bool {
+	return string(b) == s
 }
 
 func parseFloat(b *[]byte) (float64, error) {