**Description**: the Go compiler has advanced far enough where `string` and `[]byte` comparisons are fast by default. See for example https://go-review.googlesource.com/c/go/+/173323.
This PR has no impact on performance. The first commit replaces the unsafe `equalStr` with a safe version to compare benchmarks. The second commit removes `equalStr` entirely to clean up the code.

**Benchmark before change**:
```
$ go test -run nothing -bench Equal
goos: linux
goarch: amd64
pkg: github.com/grafana/jsonparser
cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics
BenchmarkEqualStr-16                            484779790                2.519 ns/op
BenchmarkBytesEqualStrSafe-16                   448411527                2.547 ns/op
BenchmarkBytesEqualStrUnsafeSlower-16           469098888                2.527 ns/op
PASS
ok      github.com/grafana/jsonparser   4.342s
```

**Benchmark after change**:
```
$ go test -run nothing -bench Equal
goos: linux
goarch: amd64
pkg: github.com/grafana/jsonparser
cpu: AMD Ryzen 7 7840U w/ Radeon  780M Graphics
BenchmarkEqualStr-16                            472333369                2.539 ns/op
BenchmarkBytesEqualStrSafe-16                   473661310                2.535 ns/op
BenchmarkBytesEqualStrUnsafeSlower-16           460124720                2.625 ns/op
PASS
ok      github.com/grafana/jsonparser   4.396s
```